# docs/swagger/swagger.yaml
definitions:
  auth.BadRequestResponse:
    properties:
      error:
        description: Error message
        example: Bad request
        type: string
    type: object
  auth.ForgotPasswordRequest:
    properties:
      email:
        description: User email address
        example: user@example.com
        type: string
    required:
    - email
    type: object
  auth.InternalServerErrorResponse:
    properties:
      error:
        description: Error message
        example: Internal server error
        type: string
    type: object
  auth.LoginRequest:
    description: Login request payload
    properties:
      email:
        description: Email for authentication (optional if username provided)
        example: admin@hotel.com
        type: string
      password:
        description: Password for authentication
        example: password123
        type: string
      tenant_id:
        example: 1
        type: integer
    required:
    - email
    - password
    type: object
  auth.LoginResponse:
    description: Login response payload
    properties:
      expired_at:
        description: Token expiration timestamp in seconds
        example: 1700000000
        type: integer
      refresh_token:
        description: JWT refresh token
        example: dGhpcyBpcyBhIHJlZnJlc2ggdG9rZW4...
        type: string
      token:
        description: JWT authentication token
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  auth.RegisterResponse:
    properties:
      created_at:
        example: "2021-01-01T00:00:00Z"
        type: string
      email:
        example: admin@hotel.com
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      is_email_verified:
        example: true
        type: boolean
      organization:
        example: org name
        type: string
      updated_at:
        example: "2021-01-01T00:00:00Z"
        type: string
    type: object
  auth.RegisterTenantRequest:
    description: Register admin request payload
    properties:
      email:
        type: string
      organization:
        minLength: 2
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - organization
    - password
    type: object
  auth.ResetAdminPasswordRequest:
    properties:
      code:
        description: Password reset code
        example: "1234567"
        type: string
      email:
        description: Admin email address
        example: admin@example.com
        type: string
      new_password:
        example: NewPassword123
        minLength: 8
        type: string
    required:
    - code
    - email
    - new_password
    type: object
  auth.UnauthorizedResponse:
    properties:
      error:
        description: Error message
        example: Unauthorized
        type: string
    type: object
  auth.VerifyEmailRequest:
    description: verify email request payload
    properties:
      code:
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
...
info:
  contact: {}
  description: |
    Nucleus ERP is an open-source, API-first business suite designed for African businesses and beyond.
    It provides a modern, modular, and developer-friendly platform for managing core business operations.
    It also provides secure, scalable, and extensible endpoints for managing users, businesses, stores, inventory, suppliers, keys, and webhooks. Features include JWT authentication, API key support, rate limiting, activity logging, and comprehensive API documentation.

    ### Key Features
    - **Authentication & Authorization**: Secure user management with JWT-based access control.
    - **Business Management**: Create and manage businesses, branches, and organizational structures.
    - **Point of Sale (POS)**: Process sales, payments, and receipts with support for multi-branch operations.
    - **Inventory Management**: Track stock levels, suppliers, purchases, and transfers.
    - **Finance & Taxation**: Manage taxes, VAT, and financial records.
    - **Extensible via Webhooks & Events**: Trigger custom workflows (e.g., when a sale or inventory update occurs).
    - **Audit Logging**: Automatic logging of key user and system activities for compliance.

    ### Target Users
    - Small to medium businesses in Africa looking for ERP solutions tailored to their workflows.
    - Developers and integrators building custom business apps on top of Nucleus API.
...

    ### Usage Notes
...

    ## Authentication
...
  title: Nucleus ERP API
...
paths:
...
  /api/v1/tenant/forgot-password:
    post:
      consumes:
      - application/json
      description: Initiate password reset by sending a reset code to the user's email
      parameters:
      - description: Forgot Password Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Reset code sent to email
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/auth.BadRequestResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/auth.UnauthorizedResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/auth.InternalServerErrorResponse'
      summary: Forgot Password
      tags:
      - Tenants
  /api/v1/tenant/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials (email)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/auth.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.UnauthorizedResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/auth.InternalServerErrorResponse'
      summary: Tenant login
      tags:
      - Tenants
  /api/v1/tenant/logout:
    post:
      consumes:
      - application/json
      description: Logout tenant and invalidate JWT token
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/auth.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.UnauthorizedResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/auth.InternalServerErrorResponse'
      summary: Tenant logout
      tags:
      - Tenants
  /api/v1/tenant/register:
    post:
      consumes:
      - application/json
      description: Create tenant with email, password and return JWT token
      parameters:
      - description: Register credentials (email and password)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Registration successful
          schema:
            $ref: '#/definitions/auth.RegisterResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/auth.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.UnauthorizedResponse'
        "500":
          description: Unable to send email at this time, request a new verification
            code for example@email.com
          schema:
            type: string
      summary: Tenant Register
      tags:
      - Tenants
  /api/v1/tenant/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password using email, reset code, and new password
      parameters:
      - description: Reset Password Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.ResetAdminPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successful
        "400":
          description: Bad request or invalid code
          schema:
            $ref: '#/definitions/auth.BadRequestResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/auth.UnauthorizedResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/auth.InternalServerErrorResponse'
      summary: Reset Password
      tags:
      - Tenants
  /api/v1/tenant/verify-email:
    post:
      consumes:
      - application/json
      description: Verify admin email with email and code
      parameters:
      - description: Verify Email Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Email verified successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/auth.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.UnauthorizedResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/auth.InternalServerErrorResponse'
      summary: Verify Tenant Email
      tags:
      - Tenants
...
securityDefinitions:
  ApiKeyAuth:
    description: 'API Key header using the ApiKey scheme. Example: "X-API-Key": ApiKey,
      "Value": {key}"'
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization:
      Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"